fixed hooks

+++ src/ray_caster/game_engine/draw_floor.c
@@ -6,7 +6,7 @@
 /*   By: awillems <awillems@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 20:08:12 by awillems          #+#    #+#             */
-/*   Updated: 2022/11/16 15:08:39 by awillems         ###   ########.fr       */
+/*   Updated: 2022/11/16 15:09:51 by awillems         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ void draw_floor(t_game *game, int x, double alpha, double heigth_drawn, double d
 	double rest_to_draw = MDDL_SCRN_HGTH - heigth_drawn + 1;
 	for (uint32_t y = 0; y < (uint32_t) rest_to_draw; y++)
 	{
-		double ratio = (1.0 + ((double) (rest_to_draw - y) / heigth_drawn));
+		double ratio = (1.0 + ((rest_to_draw - y) / heigth_drawn));
 		double floor_dist = dist / cos_a_minus_pa / ratio;
 		t_coord_f64 pos = set_f64(
 			game->player.coord.x + cosin.x * floor_dist,
@@ -62,7 +62,7 @@ void draw_floor(t_game *game, int x, double alpha, double heigth_drawn, double d
 		t_chunk *chunk = get_chunk(game, set_i32(pos.x, pos.y));
 		if (!chunk)
 			continue ;
-		draw_top_or_bottom(game, chunk->floor,		pos, set_i32(x, WIN_HEIGHT - y));
-		draw_top_or_bottom(game, chunk->ceiling,	pos, set_i32(x, y));
+		draw_top_or_bottom(game, chunk->floor, pos, set_i32(x, WIN_HEIGHT - y));
+		draw_top_or_bottom(game, chunk->ceiling, pos, set_i32(x, y));
 	}
 }
